Delete ECOSTAT-NutrientTargets.Rproj
This .Rproj file should be in each of our local repo but not sync here in the central GitHub repo. It was in the gitignore list so not sure why it was uploaded by default. testing it's removal because it seems to be creating conflicts with my pull requests in Rstudio.